+++ src/main/java/net/silentchaos512/gear/api/part/PartType.java
@@ -168,6 +168,10 @@ public final class PartType {
         return name;
     }
 
+    public String getShortName() {
+        return SilentGear.shortenId(name);
+    }
+
     public boolean isRemovable() {
         return isRemovable;
     }