+++ src/main/java/net/silentchaos512/gear/compat/jei/GearCraftingRecipeCategoryJei.java
@@ -1,5 +1,6 @@
 package net.silentchaos512.gear.compat.jei;
 
+import com.mojang.blaze3d.matrix.MatrixStack;
 import mezz.jei.api.constants.VanillaTypes;
 import mezz.jei.api.gui.IRecipeLayout;
 import mezz.jei.api.gui.drawable.IDrawable;
@@ -8,29 +9,38 @@ import mezz.jei.api.gui.ingredient.IGuiItemStackGroup;
 import mezz.jei.api.helpers.IGuiHelper;
 import mezz.jei.api.ingredients.IIngredients;
 import mezz.jei.api.recipe.category.IRecipeCategory;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.FontRenderer;
 import net.minecraft.client.resources.I18n;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.crafting.ICraftingRecipe;
+import net.minecraft.item.crafting.Ingredient;
+import net.minecraft.util.NonNullList;
 import net.minecraft.util.ResourceLocation;
+import net.minecraft.util.text.ITextComponent;
+import net.minecraft.util.text.StringTextComponent;
 import net.minecraftforge.common.crafting.IShapedRecipe;
 import net.minecraftforge.common.util.Size2i;
+import net.silentchaos512.gear.SilentGear;
+import net.silentchaos512.gear.crafting.ingredient.IGearIngredient;
 import net.silentchaos512.gear.init.ModItems;
 
 import javax.annotation.Nullable;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
 
 public class GearCraftingRecipeCategoryJei implements IRecipeCategory<ICraftingRecipe> {
-    public static final int width = 116;
-    public static final int height = 54;
+    public static final int WIDTH = 160;
+    public static final int HEIGHT = 132;
     private final IDrawable background;
     private final IDrawable icon;
     private final String localizedName;
     private final ICraftingGridHelper craftingGridHelper;
 
     public GearCraftingRecipeCategoryJei(IGuiHelper guiHelper) {
-        ResourceLocation location = new ResourceLocation("jei", "textures/gui/gui_vanilla.png");
-        this.background = guiHelper.createDrawable(location, 0, 60, 116, 54);
+        ResourceLocation location = SilentGear.getId("textures/gui/gear_crafting_jei.png");
+        this.background = guiHelper.createDrawable(location, 0, 0, WIDTH, HEIGHT);
         this.icon = guiHelper.createDrawableIngredient(new ItemStack(ModItems.BLUEPRINT_PACKAGE));
         this.localizedName = I18n.format("gui.silentgear.category.gearCrafting");
         this.craftingGridHelper = guiHelper.createCraftingGridHelper(1);
@@ -118,5 +128,36 @@ public class GearCraftingRecipeCategoryJei implements IRecipeCategory<ICraftingR
         } else {
             return null;
         }
+    }
+
+    @Override
+    public void draw(ICraftingRecipe recipe, MatrixStack matrixStack, double mouseX, double mouseY) {
+        Collection<ITextComponent> lines = new ArrayList<>();
+
+        NonNullList<Ingredient> ingredients = recipe.getIngredients();
+        for (int i = 0; i < ingredients.size(); i++) {
+            Ingredient ingredient = ingredients.get(i);
+            if (ingredient instanceof IGearIngredient) {
+                ITextComponent text = ((IGearIngredient) ingredient).getJeiHint().orElse(null);
+                if (text != null) {
+                    String prefix = (i + 1) + ": ";
+                    lines.add(new StringTextComponent(prefix).append(text));
+                }
+            }
+        }
+
+        matrixStack.push();
+        float scale = lines.size() > 5 ? 0.75f : 1f;
+        matrixStack.scale(scale, scale, 1f);
+
+        FontRenderer font = Minecraft.getInstance().fontRenderer;
+        int y = (int) (56 / scale);
+
+        for (ITextComponent line : lines) {
+            font.func_238407_a_(matrixStack, line.func_241878_f(), 0, y, -1);
+            y += 10;
+        }
+
+        matrixStack.pop();
     }
 }