+++ src/main/java/net/silentchaos512/gear/crafting/ingredient/IPartIngredient.java
@@ -1,7 +1,0 @@
-package net.silentchaos512.gear.crafting.ingredient;
-
-import net.silentchaos512.gear.api.part.PartType;
-
-public interface IPartIngredient {
-    PartType getPartType();
-}